refactor: Adapters/GUIをAdapters/IMGUIへリネーム
#124の方針決定を受け、Composition層のImmediateMode*命名との整合を取るため、
Adapters/GUI(namespace Adapters.GUI)をAdapters/IMGUI(namespace Adapters.IMGUI)へ
リネームした。#122のUIToolkit版追加時にRenderers配下が二重階層(GUI/IMGUI)に
なるのを避ける。
