+++ Assets/YukimaruGames/Terminal/Runtime/Adapters/IMGUI/Accessors/ColorPaletteAccessor.cs.meta
