+++ Assets/YukimaruGames/Terminal/Runtime/Adapters/IMGUI/Interfaces/Accessors.meta
