+++ Assets/YukimaruGames/Terminal/Runtime/Adapters/IMGUI/Interfaces/Accessors/IGUIStyleAccessor.cs
@@ -1,7 +1,7 @@
 using System;
 using UnityEngine;
 
-namespace YukimaruGames.Terminal.Adapters.GUI.Interfaces.Accessors
+namespace YukimaruGames.Terminal.Adapters.IMGUI.Interfaces.Accessors
 {
     public interface IGUIStyleAccessor :
         IGUIStyleProvider,
