+++ Assets/YukimaruGames/Terminal/Runtime/Adapters/IMGUI/Renderers/ClipboardRenderer.cs.meta
