+++ Assets/YukimaruGames/Terminal/Runtime/Composition/Installer/ImmediateModeInstaller.cs
@@ -15,9 +15,9 @@ using System.Linq;
 using System.Reflection;
 using System.Threading.Tasks;
 using UnityEngine;
-using YukimaruGames.Terminal.Adapters.GUI;
-using YukimaruGames.Terminal.Adapters.GUI.Accessors;
-using YukimaruGames.Terminal.Adapters.GUI.Renderers;
+using YukimaruGames.Terminal.Adapters.IMGUI;
+using YukimaruGames.Terminal.Adapters.IMGUI.Accessors;
+using YukimaruGames.Terminal.Adapters.IMGUI.Renderers;
 using YukimaruGames.Terminal.Application.Interfaces;
 using YukimaruGames.Terminal.Application.Services;
 using YukimaruGames.Terminal.Domain.Contracts.Interfaces.Repositories;
@@ -31,14 +31,14 @@ using YukimaruGames.Terminal.Infrastructure.Discoverer;
 using YukimaruGames.Terminal.Infrastructure.Factories;
 using YukimaruGames.Terminal.Infrastructure.Modes;
 using YukimaruGames.Terminal.Infrastructure.Repositories;
-using YukimaruGames.Terminal.Adapters.GUI.Coordinators;
+using YukimaruGames.Terminal.Adapters.IMGUI.Coordinators;
 using YukimaruGames.Terminal.Presentation.Accessors;
 using YukimaruGames.Terminal.Presentation.Animators;
 using YukimaruGames.Terminal.Presentation.Constants;
 using YukimaruGames.Terminal.Presentation.Contracts;
 using YukimaruGames.Terminal.Presentation.Coordinators;
 using YukimaruGames.Terminal.Presentation.Events;
-using YukimaruGames.Terminal.Adapters.GUI.Interfaces.Accessors;
+using YukimaruGames.Terminal.Adapters.IMGUI.Interfaces.Accessors;
 using YukimaruGames.Terminal.Presentation.Interfaces.Accessors;
 using YukimaruGames.Terminal.Presentation.Interfaces.Accessors.Window;
 using YukimaruGames.Terminal.Presentation.Interfaces.Coordinators;
